review: fix test and schema

+++ database/migrations/20210105124948-product.js
@@ -27,12 +27,12 @@ module.exports = {
         defaultValue: 5
       },
       buyPrice: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.INTEGER,
         allowNull: false,
         defaultValue: 0
       },
       salePrice: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.INTEGER,
         allowNull: false,
         defaultValue: 0
       },
